batch: Move realized provenance into package
The realization package owns entry views while their compact source and
target provenance remains at the batch package root.

Entry storage relies on provenance to retain where assembled lines
originated, so separating those values obscures one representation
boundary.

This commit moves realized provenance beside the entry views and keeps the
selected import path temporarily available.

Compact realized-entry storage will move next and consume both nested
modules.
